+++ src/complex_native/c32_impl.rs
@@ -1074,3 +1074,14 @@ impl Distribution<c32> for StandardNormal {
         }
     }
 }
+
+impl core::iter::Sum for c32 {
+    fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
+        iter.fold(c32::new(0.0, 0.0), |acc, e| acc + e)
+    }
+}
+impl core::iter::Product for c32 {
+    fn product<I: Iterator<Item = Self>>(iter: I) -> Self {
+        iter.fold(c32::new(0.0, 0.0), |acc, e| acc * e)
+    }
+}